+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.2.0 -- 2022-11-18
 * `breakpoint` includes a `*result` binding in its output
+* Fix a bug breaking Windows compatibility
 
 ## 0.1.1.1 -- 2022-11-02
 * Support `IsString` version of string literals in `excludeVars`